Vending Machine & Dialogue changes
Created and coded the vending machine script. The player can interact with a vending machine once and it will heal them for 10 HP (will probably change to full health).

Changed it so that when the character enters box collider instead of dialogue playing instantly they must also now hit space bar to activate.