Delete Ghost object if combat is won
When the player enters combat the gameObject that they touched to bring them to the combat scene is saved. If the player wins combat the name is then sent to a function in gameManager. In this function the name is saved into an array. Once the player returns to the scene the SceneLoader loops through that array destroying all the ghost objects that have previously been defeated!