Added UI, Bloom effect, and Dissolve material
I added a bloom effect on the main camera in act 1 to make it pop.

I created the UI for the lab where the ghosts that you capture show up and you can select their images and on the right side an overview will pop up showing their information. There is the area where you will drag and drop party members *WIP*

I also made a custom material that dissolves when the character enters the range of the object.

Upload some 2D assets as well.